+++ plugins/workload-management/src/main/java/org/opensearch/plugin/wlm/WorkloadManagementPlugin.java
@@ -9,7 +9,6 @@
 package org.opensearch.plugin.wlm;
 
 import org.opensearch.action.ActionRequest;
-import org.opensearch.action.support.ActionFilter;
 import org.opensearch.cluster.metadata.IndexNameExpressionResolver;
 import org.opensearch.cluster.node.DiscoveryNodes;
 import org.opensearch.cluster.service.ClusterService;
@@ -38,7 +37,7 @@ import org.opensearch.plugin.wlm.rest.RestDeleteWorkloadGroupAction;
 import org.opensearch.plugin.wlm.rest.RestGetWorkloadGroupAction;
 import org.opensearch.plugin.wlm.rest.RestUpdateWorkloadGroupAction;
 import org.opensearch.plugin.wlm.rule.WorkloadGroupFeatureType;
-import org.opensearch.plugin.wlm.rule.sync.RefreshBasedSyncMechanism;
+import org.opensearch.plugin.wlm.rule.WorkloadGroupFeatureValueValidator;
 import org.opensearch.plugin.wlm.service.WorkloadGroupPersistenceService;
 import org.opensearch.plugins.ActionPlugin;
 import org.opensearch.plugins.Plugin;
@@ -46,14 +45,10 @@ import org.opensearch.plugins.SystemIndexPlugin;
 import org.opensearch.repositories.RepositoriesService;
 import org.opensearch.rest.RestController;
 import org.opensearch.rest.RestHandler;
-import org.opensearch.rule.InMemoryRuleProcessingService;
-import org.opensearch.rule.RuleEntityParser;
 import org.opensearch.rule.RulePersistenceService;
 import org.opensearch.rule.autotagging.FeatureType;
 import org.opensearch.rule.service.IndexStoredRulePersistenceService;
 import org.opensearch.rule.spi.RuleFrameworkExtension;
-import org.opensearch.rule.storage.AttributeValueStoreFactory;
-import org.opensearch.rule.storage.DefaultAttributeValueStore;
 import org.opensearch.rule.storage.IndexBasedRuleQueryMapper;
 import org.opensearch.rule.storage.XContentRuleParser;
 import org.opensearch.script.ScriptService;
@@ -62,6 +57,7 @@ import org.opensearch.transport.client.Client;
 import org.opensearch.watcher.ResourceWatcherService;
 
 import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
 import java.util.function.Supplier;
 
@@ -69,6 +65,7 @@ import java.util.function.Supplier;
  * Plugin class for WorkloadManagement
  */
 public class WorkloadManagementPlugin extends Plugin implements ActionPlugin, SystemIndexPlugin, RuleFrameworkExtension {
+
     /**
      * The name of the index where rules are stored.
      */
@@ -79,8 +76,6 @@ public class WorkloadManagementPlugin extends Plugin implements ActionPlugin, Sy
     public static final int MAX_RULES_PER_PAGE = 50;
 
     private final RulePersistenceServiceHolder rulePersistenceServiceHolder = new RulePersistenceServiceHolder();
-
-    private AutoTaggingActionFilter autoTaggingActionFilter;
 
     /**
      * Default constructor
@@ -101,36 +96,16 @@ public class WorkloadManagementPlugin extends Plugin implements ActionPlugin, Sy
         IndexNameExpressionResolver indexNameExpressionResolver,
         Supplier<RepositoriesService> repositoriesServiceSupplier
     ) {
-        RuleEntityParser parser = new XContentRuleParser(WorkloadGroupFeatureType.INSTANCE);
-        AttributeValueStoreFactory attributeValueStoreFactory = new AttributeValueStoreFactory(
-            WorkloadGroupFeatureType.INSTANCE,
-            DefaultAttributeValueStore::new
-        );
-        InMemoryRuleProcessingService ruleProcessingService = new InMemoryRuleProcessingService(attributeValueStoreFactory);
+        FeatureTypeHolder.featureType = new WorkloadGroupFeatureType(new WorkloadGroupFeatureValueValidator(clusterService));
         RulePersistenceServiceHolder.rulePersistenceService = new IndexStoredRulePersistenceService(
             INDEX_NAME,
             client,
+            clusterService,
             MAX_RULES_PER_PAGE,
-            parser,
+            new XContentRuleParser(FeatureTypeHolder.featureType),
             new IndexBasedRuleQueryMapper()
         );
-
-        RefreshBasedSyncMechanism refreshMechanism = new RefreshBasedSyncMechanism(
-            client,
-            threadPool,
-            clusterService.getSettings(),
-            clusterService.getClusterSettings(),
-            parser,
-            ruleProcessingService,
-            WorkloadGroupFeatureType.INSTANCE
-        );
-
-        return List.of(refreshMechanism);
-    }
-
-    @Override
-    public List<ActionFilter> getActionFilters() {
-        return List.of(autoTaggingActionFilter);
+        return Collections.emptyList();
     }
 
     @Override
@@ -168,7 +143,7 @@ public class WorkloadManagementPlugin extends Plugin implements ActionPlugin, Sy
 
     @Override
     public List<Setting<?>> getSettings() {
-        return List.of(WorkloadGroupPersistenceService.MAX_QUERY_GROUP_COUNT, RefreshBasedSyncMechanism.RULE_SYNC_REFRESH_INTERVAL_SETTING);
+        return List.of(WorkloadGroupPersistenceService.MAX_QUERY_GROUP_COUNT);
     }
 
     @Override
@@ -183,10 +158,14 @@ public class WorkloadManagementPlugin extends Plugin implements ActionPlugin, Sy
 
     @Override
     public FeatureType getFeatureType() {
-        return WorkloadGroupFeatureType.INSTANCE;
+        return FeatureTypeHolder.featureType;
     }
 
     static class RulePersistenceServiceHolder {
         private static RulePersistenceService rulePersistenceService;
+    }
+
+    static class FeatureTypeHolder {
+        private static FeatureType featureType;
     }
 }